refactor: rename factories
