+++ packages/platform-sdk-btc-server/source/index.ts
@@ -3,7 +3,7 @@ import Hapi from "@hapi/hapi";
 import Joi from "joi";
 import { v4 as uuidv4 } from "uuid";
 
-import { useClient, useDatabase, useLogger } from "./helpers";
+import { useClient, useDatabase, useLogger } from "./factories";
 
 export const subscribe = async (flags: {
 	coin: string;
