+++ packages/platform-sdk-egld/source/public-key.service.ts
@@ -1,6 +1,6 @@
 import { IoC, Services } from "@arkecosystem/platform-sdk";
 
-import { makeAccount } from "./helpers";
+import { makeAccount } from "./factories";
 
 @IoC.injectable()
 export class PublicKeyService extends Services.AbstractPublicKeyService {
