+++ packages/platform-sdk-news-indexer/source/factories.ts
